requirements.txt
PyYAML==3.10
boto==2.3.0
